requirements.txt
python-telegram-bot==21.7
psutil==6.1.1
python-dotenv==1.0.1
aiofiles==24.1.0
